demo/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
pydantic==2.5.2
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
spacy==3.7.2
transformers==4.35.2
torch==2.1.1
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2 